+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Web UI version numbers should always match the corresponding version of LBRY App
   * You can now make API calls directly on the lbry module, e.g. lbry.peer_list()
   * New-style API calls return promises instead of using callbacks
   * Wherever possible, use outpoints for unique IDs instead of names or SD hashes
+  * New publishes now display immediately in My Files, even before they hit the lbrynet file manager.
 
 ### Changed
   * Update process now easier and more reliable